ST Port: Make STV player faster.

+++ tiberiandawn/atarilib/stvq/stvq_io.h
@@ -12,8 +12,11 @@ extern "C" {
 
 typedef struct StvqIo {
 	void *user;
-	/* 0 ok, -1 err; must transfer exactly n bytes. */
-	int (*read)(void *user, void *buf, size_t n);
+	/*
+	 * Read up to n bytes; return count transferred.
+	 * Short read means EOF (not a hard error).
+	 */
+	size_t (*read)(void *user, void *buf, size_t n);
 	/* 0 ok, -1 err; whence is SEEK_SET / SEEK_CUR / SEEK_END. */
 	int (*seek)(void *user, long off, int whence);
 } StvqIo;
